+++ backtrack/app/Http/Controllers/TabController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Tab;
 use App\Track;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TabController extends Controller
 {
@@ -15,6 +16,9 @@ class TabController extends Controller
     public function save(Request $request) {
         $json = ["success"=>true, "message"=>"Tab saved"];
         try {
+            if(!Auth::user()->super) {
+                throw new \Exception("You cannot add tabs");
+            }
             $row = [
                 "song_id" => $request->get("song_id"),
                 "instrument" => $request->get("instrument"),
@@ -28,7 +32,7 @@ class TabController extends Controller
             }
             Tab::updateOrCreate(['id'=>$row['id']],$row);
         } catch (\Exception $e) {
-            $json['message'] = $e->getMessage()." ".$e->getFile()." ".$e->getLine();
+            $json['message'] = $e->getMessage();
             $json['success'] = false;
         }
         return json_encode($json);
@@ -48,6 +52,9 @@ class TabController extends Controller
     public function delete($id, Request $request) {
         $json = ['success'=>true];
         try {
+            if(!Auth::user()->super) {
+                throw new \Exception("You cannot add tabs");
+            }
             $tab = Tab::findOrFail($id);
             $tab->delete();
         } catch (\Exception $e) {
